Cleanup platform component implementation.
