--- core/database/src/commonMain/kotlin/com.thomaskioko.tvmaniac.db/DatabaseComponent.kt
@@ -1,3 +1,0 @@
-package com.thomaskioko.tvmaniac.db
-
-expect interface DatabaseComponent
