+++ core/database/src/commonMain/kotlin/com.thomaskioko.tvmaniac.db/DatabasePlatformComponent.kt
@@ -1,8 +1,6 @@
 package com.thomaskioko.tvmaniac.db
 
-import android.app.Application
 import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.driver.android.AndroidSqliteDriver
 import com.thomaskioko.tvmaniac.core.db.Episode
 import com.thomaskioko.tvmaniac.core.db.Episode_image
 import com.thomaskioko.tvmaniac.core.db.Last_requests
@@ -17,18 +15,10 @@ import com.thomaskioko.tvmaniac.core.db.Watchlist
 import com.thomaskioko.tvmaniac.util.scope.ApplicationScope
 import me.tatarka.inject.annotations.Provides
 
-actual interface DatabaseComponent {
+expect interface DatabasePlatformComponent
 
-    @ApplicationScope
-    @Provides
-    fun provideSqlDriver(
-        application: Application,
-    ): SqlDriver = AndroidSqliteDriver(
-        schema = TvManiacDatabase.Schema,
-        context = application,
-        name = "tvShows.db",
-    )
 
+interface DatabaseComponent : DatabasePlatformComponent {
     @ApplicationScope
     @Provides
     fun provideTvManiacDatabase(
