--- core/database/src/iosMain/kotlin/com/thomaskioko/tvmaniac/db/DatabaseComponent.kt
@@ -1,68 +1,0 @@
-package com.thomaskioko.tvmaniac.db
-
-import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.driver.native.NativeSqliteDriver
-import com.thomaskioko.tvmaniac.core.db.Episode
-import com.thomaskioko.tvmaniac.core.db.Episode_image
-import com.thomaskioko.tvmaniac.core.db.Last_requests
-import com.thomaskioko.tvmaniac.core.db.Season
-import com.thomaskioko.tvmaniac.core.db.Show
-import com.thomaskioko.tvmaniac.core.db.Show_category
-import com.thomaskioko.tvmaniac.core.db.Show_image
-import com.thomaskioko.tvmaniac.core.db.Similar_shows
-import com.thomaskioko.tvmaniac.core.db.Trailers
-import com.thomaskioko.tvmaniac.core.db.TvManiacDatabase
-import com.thomaskioko.tvmaniac.core.db.Watchlist
-import com.thomaskioko.tvmaniac.util.scope.ApplicationScope
-import me.tatarka.inject.annotations.Provides
-
-actual interface DatabaseComponent {
-
-    @ApplicationScope
-    @Provides
-    fun provideSqlDriver(): SqlDriver = NativeSqliteDriver(TvManiacDatabase.Schema, "tvShows.db")
-
-    @ApplicationScope
-    @Provides
-    fun provideTvManiacDatabase(
-        sqlDriver: SqlDriver,
-    ): TvManiacDatabase = TvManiacDatabase(
-        driver = sqlDriver,
-        showAdapter = Show.Adapter(
-            genresAdapter = stringColumnAdapter,
-            idAdapter = IdAdapter(),
-        ),
-        last_requestsAdapter = Last_requests.Adapter(
-            timestampAdapter = InstantColumnAdapter,
-        ),
-        episode_imageAdapter = Episode_image.Adapter(
-            idAdapter = IdAdapter(),
-            tmdb_idAdapter = IdAdapter(),
-        ),
-        episodeAdapter = Episode.Adapter(
-            idAdapter = IdAdapter(),
-            season_idAdapter = IdAdapter(),
-        ),
-        seasonAdapter = Season.Adapter(
-            idAdapter = IdAdapter(),
-            show_idAdapter = IdAdapter(),
-        ),
-        show_imageAdapter = Show_image.Adapter(
-            idAdapter = IdAdapter(),
-        ),
-        similar_showsAdapter = Similar_shows.Adapter(
-            idAdapter = IdAdapter(),
-            similar_show_idAdapter = IdAdapter(),
-        ),
-        watchlistAdapter = Watchlist.Adapter(
-            idAdapter = IdAdapter(),
-        ),
-        show_categoryAdapter = Show_category.Adapter(
-            idAdapter = IdAdapter(),
-            category_idAdapter = IdAdapter(),
-        ),
-        trailersAdapter = Trailers.Adapter(
-            show_idAdapter = IdAdapter(),
-        ),
-    )
-}
