+++ core/database/src/iosMain/kotlin/com/thomaskioko/tvmaniac/db/DatabasePlatformComponent.kt
@@ -1,0 +1,25 @@
+package com.thomaskioko.tvmaniac.db
+
+import app.cash.sqldelight.db.SqlDriver
+import app.cash.sqldelight.driver.native.NativeSqliteDriver
+import com.thomaskioko.tvmaniac.core.db.Episode
+import com.thomaskioko.tvmaniac.core.db.Episode_image
+import com.thomaskioko.tvmaniac.core.db.Last_requests
+import com.thomaskioko.tvmaniac.core.db.Season
+import com.thomaskioko.tvmaniac.core.db.Show
+import com.thomaskioko.tvmaniac.core.db.Show_category
+import com.thomaskioko.tvmaniac.core.db.Show_image
+import com.thomaskioko.tvmaniac.core.db.Similar_shows
+import com.thomaskioko.tvmaniac.core.db.Trailers
+import com.thomaskioko.tvmaniac.core.db.TvManiacDatabase
+import com.thomaskioko.tvmaniac.core.db.Watchlist
+import com.thomaskioko.tvmaniac.util.scope.ApplicationScope
+import me.tatarka.inject.annotations.Provides
+
+actual interface DatabasePlatformComponent {
+
+    @ApplicationScope
+    @Provides
+    fun provideSqlDriver(): SqlDriver = NativeSqliteDriver(TvManiacDatabase.Schema, "tvShows.db")
+
+}
