--- core/database/src/jvmMain/kotlin/com/thomaskioko/tvmaniac/db/DatabaseComponent.kt
@@ -1,3 +1,0 @@
-package com.thomaskioko.tvmaniac.db
-
-actual interface DatabaseComponent
