+++ core/database/src/jvmMain/kotlin/com/thomaskioko/tvmaniac/db/DatabasePlatformComponent.kt
@@ -1,0 +1,3 @@
+package com.thomaskioko.tvmaniac.db
+
+actual interface DatabasePlatformComponent
