+++ core/datastore/implementation/src/androidMain/kotlin/com/thomaskioko/tvmaniac/datastore/implementation/DataStorePlatformComponent.kt
@@ -20,7 +20,4 @@ actual interface DataStorePlatformComponent {
         producePath = { context.filesDir.resolve(dataStoreFileName).absolutePath },
     )
 
-    @ApplicationScope
-    @Provides
-    fun provideDatastoreRepository(bind: DatastoreRepositoryImpl): DatastoreRepository = bind
 }
