+++ core/datastore/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/datastore/implementation/DataStorePlatformComponent.kt
@@ -1,3 +1,15 @@
 package com.thomaskioko.tvmaniac.datastore.implementation
 
+import com.thomaskioko.tvmaniac.datastore.api.DatastoreRepository
+import com.thomaskioko.tvmaniac.util.scope.ApplicationScope
+import me.tatarka.inject.annotations.Provides
+
 expect interface DataStorePlatformComponent
+
+
+interface DataStoreComponent : DataStorePlatformComponent {
+
+    @ApplicationScope
+    @Provides
+    fun provideDatastoreRepository(bind: DatastoreRepositoryImpl): DatastoreRepository = bind
+}
