+++ core/datastore/implementation/src/iosMain/kotlin/com/thomaskioko/tvmaniac/datastore/implementation/DataStorePlatformComponent.kt
@@ -33,7 +33,4 @@ actual interface DataStorePlatformComponent {
         },
     )
 
-    @ApplicationScope
-    @Provides
-    fun provideDatastoreRepository(bind: DatastoreRepositoryImpl): DatastoreRepository = bind
 }
