+++ core/trakt-api/implementation/src/commonMain/kotlin/com/thomaskioko/trakt/service/implementation/inject/TraktComponent.kt
@@ -21,7 +21,7 @@ import me.tatarka.inject.annotations.Provides
 typealias TraktHttpClient = HttpClient
 typealias TraktJson = Json
 
-interface TraktComponent {
+interface TraktComponent : TraktPlatformComponent {
 
     @ApplicationScope
     @Provides
