+++ shared/src/iosMain/kotlin/com.thomaskioko.tvmaniac.shared/base/ApplicationComponent.kt
@@ -1,10 +1,9 @@
 package com.thomaskioko.tvmaniac.shared.base
 
 import com.thomaskioko.trakt.service.implementation.inject.TraktComponent
-import com.thomaskioko.trakt.service.implementation.inject.TraktPlatformComponent
 import com.thomaskioko.tvmaniac.data.category.implementation.CategoryComponent
 import com.thomaskioko.tvmaniac.data.trailers.implementation.TrailerComponent
-import com.thomaskioko.tvmaniac.datastore.implementation.DataStorePlatformComponent
+import com.thomaskioko.tvmaniac.datastore.implementation.DataStoreComponent
 import com.thomaskioko.tvmaniac.db.DatabaseComponent
 import com.thomaskioko.tvmaniac.episodeimages.implementation.EpisodeImageComponent
 import com.thomaskioko.tvmaniac.episodes.implementation.EpisodeComponent
@@ -24,7 +23,6 @@ import com.thomaskioko.tvmaniac.showimages.implementation.ShowImagesComponent
 import com.thomaskioko.tvmaniac.shows.implementation.DiscoverComponent
 import com.thomaskioko.tvmaniac.similar.implementation.SimilarShowsComponent
 import com.thomaskioko.tvmaniac.tmdb.implementation.TmdbComponent
-import com.thomaskioko.tvmaniac.tmdb.implementation.TmdbPlatformComponent
 import com.thomaskioko.tvmaniac.traktauth.implementation.TraktAuthenticationComponent
 import com.thomaskioko.tvmaniac.util.inject.UtilPlatformComponent
 import com.thomaskioko.tvmaniac.util.scope.ApplicationScope
@@ -36,7 +34,7 @@ import me.tatarka.inject.annotations.Component
 abstract class ApplicationComponent :
     CategoryComponent,
     DatabaseComponent,
-    DataStorePlatformComponent,
+    DataStoreComponent,
     EpisodeComponent,
     EpisodeImageComponent,
     ProfileComponent,
@@ -48,9 +46,7 @@ abstract class ApplicationComponent :
     SimilarShowsComponent,
     StatsComponent,
     TmdbComponent,
-    TmdbPlatformComponent,
     TraktComponent,
-    TraktPlatformComponent,
     TraktAuthenticationComponent,
     TrailerComponent,
     UtilPlatformComponent,
